+++ include/localization/gettext.h
@@ -23,12 +23,14 @@
 #ifndef GETTEXT_H
 #define GETTEXT_H
 
+#include <format>
 #include <string>
 #include <libintl.h>
 
 #define GETTEXT_CONTEXT_SEPARATOR "\004"
 #define _(String) dgettext(::Nickvision::Localization::Gettext::getDomainName().c_str(), String)
 #define _n(String, StringPlural, N) dngettext(::Nickvision::Localization::Gettext::getDomainName().c_str(), String, StringPlural, static_cast<unsigned long>(N))
+#define _f(String, ...) ::Nickvision::Localization::Gettext::fgettext(String, __VA_ARGS__)
 #define _p(Context, String) ::Nickvision::Localization::Gettext::pgettext(Context GETTEXT_CONTEXT_SEPARATOR String, String)
 #define _pn(Context, String, StringPlural, N) ::Nickvision::Localization::Gettext::pngettext(Context GETTEXT_CONTEXT_SEPARATOR String, String, StringPlural, static_cast<unsigned long>(N))
 
@@ -45,6 +47,18 @@ namespace Nickvision::Localization::Gettext
      * @return The gettext domain name
      */
     const std::string& getDomainName();
+    /**
+     * @brief Translates a message and formats it with the given arguments.
+     * @param msg The message to translate
+     * @param args The arguments to format the translated message with
+     */
+    template<typename... Args>
+    const char* fgettext(const char* msg, Args&&... args)
+    {
+        static std::string res;
+        res = std::vformat(_(msg), std::make_format_args(args...));
+        return res.c_str();
+    }
     /**
      * @brief Translates a message for a given context.
      * @param context The context of the message